requirements.txt
discord.py
python-dotenv
yfpy
table2ascii
selenium


